Swagger codegen! API driven development!
This commit generates code from the swagger/swagger.json.

To make this happen with dropwizard some trickery was needed. The
swagger-cli/templates/api.mustache is our custom template that creates a
dropwizard-friendly resource with @Auth annotations in the methods. It
shouldn't be necessary, but if other templates need to be overridden, they can
be found from
https://github.com/swagger-api/swagger-codegen/tree/v2.3.1/modules/swagger-codegen/src/main/resources/JavaJaxRS.

generate-api.sh generates the V1Api, copies it to the relevant src directory
and deletes it thereafter.

PROs:
- @ApiModelProperty doesn't need to be on the model classes.
- Got rid of @post, @path("/v1/payments"), @Consumes, @Produces, @ApiOperation,
  @ApiResponses and @ApiResponse annotations on the old Resource classes.
- We have API driven development, which means a technical author or a product
  manager can add/remove features from the swagger.json, create a PR which will
  then break the build. The developer then makes the necessary changes.

CONs:
- One swagger file generates one REST java class. This could become bloated. On
  the other hand, a pro might be that it forces us to keep our microservices
simple.
- Unknown unknowns... what if we need some dropwizard features now or down the
  line which aren't easily editable in the api.mustache templates? At the
  moment we are saved by the fact that every method needs an @Auth annotation for
  instance, but what happens if only one method needs some particular annotation
  later on?

@oswaldquek
