+++ .gitignore
@@ -30,3 +30,5 @@ target
 bin
 
 .DS_Store
+
+src/main/java/uk/gov/pay/api/resources/V1Api.java
