+++ swagger-cli/templates/api.mustache
@@ -1,0 +1,67 @@
+package {{package}};
+
+import io.dropwizard.auth.Auth;
+import uk.gov.pay.api.auth.Account;
+import uk.gov.pay.api.resources.PaymentsResource;
+
+import javax.inject.Inject;
+
+import io.swagger.annotations.ApiParam;
+import io.swagger.jaxrs.*;
+
+{{#imports}}import {{import}};
+{{/imports}}
+
+import java.util.Map;
+import java.util.List;
+
+import java.io.InputStream;
+
+import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
+import org.glassfish.jersey.media.multipart.FormDataParam;
+
+import javax.servlet.ServletConfig;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+import javax.ws.rs.*;
+{{#useBeanValidation}}
+    import javax.validation.constraints.*;
+{{/useBeanValidation}}
+
+@Path("/{{{baseName}}}")
+{{#hasConsumes}}@Consumes({ {{#consumes}}"{{{mediaType}}}"{{#hasMore}}, {{/hasMore}}{{/consumes}} }){{/hasConsumes}}
+{{#hasProduces}}@Produces({ {{#produces}}"{{{mediaType}}}"{{#hasMore}}, {{/hasMore}}{{/produces}} }){{/hasProduces}}
+@io.swagger.annotations.Api(description = "the {{{baseName}}} API")
+{{>generatedAnnotation}}
+{{#operations}}
+    public class {{classname}}  {
+    private final PaymentsResource delegate;
+
+    @Inject
+    public {{classname}}(PaymentsResource delegate) {
+    this.delegate = delegate;
+    }
+
+    {{#operation}}
+        @{{httpMethod}}
+        {{#subresourceOperation}}@Path("{{{path}}}"){{/subresourceOperation}}
+        {{#hasConsumes}}@Consumes({ {{#consumes}}"{{{mediaType}}}"{{#hasMore}}, {{/hasMore}}{{/consumes}} }){{/hasConsumes}}
+        {{#hasProduces}}@Produces({ {{#produces}}"{{{mediaType}}}"{{#hasMore}}, {{/hasMore}}{{/produces}} }){{/hasProduces}}
+        @io.swagger.annotations.ApiOperation(value = "{{{summary}}}", notes = "{{{notes}}}"{{#returnBaseType}}, response = {{{returnBaseType}}}.class{{/returnBaseType}}{{#returnContainer}}, responseContainer = "{{{returnContainer}}}"{{/returnContainer}}{{#hasAuthMethods}}, authorizations = {
+        {{#authMethods}}@io.swagger.annotations.Authorization(value = "{{name}}"{{#isOAuth}}, scopes = {
+        {{#scopes}}@io.swagger.annotations.AuthorizationScope(scope = "{{scope}}", description = "{{description}}"){{#hasMore}},
+        {{/hasMore}}{{/scopes}}
+        }{{/isOAuth}}){{#hasMore}},
+        {{/hasMore}}{{/authMethods}}
+        }{{/hasAuthMethods}}, tags={ {{#vendorExtensions.x-tags}}"{{tag}}",{{/vendorExtensions.x-tags}} })
+        @io.swagger.annotations.ApiResponses(value = { {{#responses}}
+            @io.swagger.annotations.ApiResponse(code = {{{code}}}, message = "{{{message}}}"{{#baseType}}, response = {{{baseType}}}.class{{/baseType}}{{#containerType}}, responseContainer = "{{{containerType}}}"{{/containerType}}){{#hasMore}},
+            {{/hasMore}}{{/responses}} })
+        public Response {{nickname}}({{#allParams}}{{>queryParams}}{{>pathParams}}{{>headerParams}}{{>bodyParams}}{{>formParams}},{{/allParams}}@ApiParam(value = "accountId", hidden = true) @Auth Account account)
+        {
+        return delegate.{{nickname}}({{#allParams}}{{#isFile}}{{paramName}}InputStream, {{paramName}}Detail{{/isFile}}{{^isFile}}{{paramName}}{{/isFile}},{{/allParams}}account);
+        }
+    {{/operation}}
+    }
+{{/operations}}
