+++ swagger/swagger.json
@@ -119,7 +119,7 @@
       "post" : {
         "summary" : "Create new payment",
         "description" : "Create a new payment for the account associated to the Authorisation token. The Authorisation token needs to be specified in the 'authorization' header as 'authorization: Bearer YOUR_API_KEY_HERE'",
-        "operationId" : "newPayment",
+        "operationId" : "createNewPayment",
         "consumes" : [ "application/json" ],
         "produces" : [ "application/json" ],
         "parameters" : [ {
